requirements.txt
numpy
pillow
streamlit
tensorflow
h5py
huggingface_hub
python-dotenv
